--- a/lib/Cargo.toml
+++ b/lib/Cargo.toml
@@ -151,3 +151,7 @@ harness = false
 [[bench]]
 name = "processor"
 harness = false
+
+[[bench]]
+name = "index_btree"
+harness = false