--- a/lib/src/doc/index.rs
+++ b/lib/src/doc/index.rs
@@ -3,6 +3,7 @@ use crate::dbs::Statement;
 use crate::dbs::{Options, Transaction};
 use crate::doc::{CursorDoc, Document};
 use crate::err::Error;
+use crate::idx::btree::store::BTreeStoreType;
 use crate::idx::ft::FtIndex;
 use crate::idx::IndexKeyBase;
 use crate::sql::array::Array;
@@ -185,12 +186,13 @@ impl<'a> IndexOperation<'a> {
 	) -> Result<(), Error> {
 		let ikb = IndexKeyBase::new(self.opt, self.ix);
 		let az = run.get_az(self.opt.ns(), self.opt.db(), az.as_str()).await?;
-		let mut ft = FtIndex::new(run, az, ikb, order, scoring, hl).await?;
+		let mut ft = FtIndex::new(run, az, ikb, order, scoring, hl, BTreeStoreType::Write).await?;
 		if let Some(n) = &self.n {
 			// TODO: Apply the analyzer
-			ft.index_document(run, self.rid, n).await
+			ft.index_document(run, self.rid, n).await?;
 		} else {
-			ft.remove_document(run, self.rid).await
+			ft.remove_document(run, self.rid).await?;
 		}
+		ft.finish(run).await
 	}
 }