… Cargo.lock
… lib/src/idx/ft/docids.rs
… lib/src/idx/ft/doclength.rs
… lib/src/idx/ft/mod.rs
… lib/src/idx/ft/postings.rs
… lib/src/idx/ft/scorer.rs
… lib/src/idx/ft/terms.rs
… lib/src/idx/planner/executor.rs
… lib/src/idx/planner/plan.rs
… lib/src/sql/statements/analyze.rs