## What is the motivation?

Our initial implementation of full-text indexing was primarily focused on achieving correctness. The next step is to ensure optimal performance by implementing every possible optimization.

### BTree nodes

In the existing system, when a document is indexed, the underlying BTree structure can be updated multiple times. For instance, each new term creates a new entry in the term BTree.

Currently, the initial implementation serializes and deserializes a node each time it is read (during BTree traversal) and updated (during BTree insertion and deletion). This can happen multiple times for a single document.

### FST keys

We have two types of BTrees at the moment: one with keys stored in a FST (for terms) and the other with keys stored in a Trie (for all other instances). We've observed that updating keys in a FST can be quite slow, as it has to be reconstructed.

## What does this change do?

This PR introduces two optimisations:

###  new layer on top of the BTree

 There is 3 types:

1. **Read**: This layer functions as a long-term cache, using an LRU map to store deserialized nodes. This cache is typically used during large iterations in search operations. We aim to store the most frequently accessed nodes (usually at the top of the BTree) rather than all nodes.
2. **Write**: This layer acts as a short-term cache, storing all nodes in memory. It also defers serialization to the KVStore, which happens just before a transaction is committed. This cache never holds a large number of nodes, only those required for indexing the current documents in the current transaction.
3. **Traversal**: This layer maintains the previous behavior. It is primarily used when the entire BTree needs to be traversed, typically for statistical purposes.

### FST keys optimisation

When the BTree is updated, multiple nodes are typically modified, and it's common for the same nodes to be altered several times. When keys in the FST are changed, they are converted into a Trie format. Consequently, any further modifications can be made very quickly. The FST is only rebuilt when the node is serialized, which occurs when the transaction is committed.

## What is your testing strategy?

Test has been updated, and a bench called `index_btree` has been created.

```bash
cargo bench --package surrealdb --no-default-features --features kv-mem,scripting,http index_btree
```

The bench before the optimisations:

```
     Running benches/index_btree.rs (target/release/deps/index_btree-4ec0233e1cfcfba6)
index_btree/btree-insertion-fst
                        time:   [1.1601 s 1.1607 s 1.1613 s]
                        thrpt:  [0.8611  elem/s 0.8615  elem/s 0.8620  elem/s]
                 change:
                        time:   [-0.2121% -0.1250% -0.0409%] (p = 0.01 < 0.05)
                        thrpt:  [+0.0409% +0.1252% +0.2125%]
                        Change within noise threshold.
index_btree/btree-insertion-trie
                        time:   [2.1609 s 2.1848 s 2.2104 s]
                        thrpt:  [0.4524  elem/s 0.4577  elem/s 0.4628  elem/s]
                 change:
                        time:   [-7.4256% -4.6424% -1.5198%] (p = 0.01 < 0.05)
                        thrpt:  [+1.5433% +4.8684% +8.0212%]
                        Performance has improved.
```

The same bench after:

```
     Running benches/index_btree.rs (target/release/deps/index_btree-9597b6ae09b532cb)
index_btree/btree-insertion-fst
                        time:   [78.270 ms 78.347 ms 78.435 ms]
                        thrpt:  [12.749  elem/s 12.764  elem/s 12.776  elem/s]
                 change:
                        time:   [-93.256% -93.247% -93.238%] (p = 0.00 < 0.05)
                        thrpt:  [+1378.9% +1380.9% +1382.9%]
                        Performance has improved.
index_btree/btree-insertion-trie
                        time:   [78.301 ms 78.438 ms 78.549 ms]
                        thrpt:  [12.731  elem/s 12.749  elem/s 12.771  elem/s]
                 change:
                        time:   [-96.453% -96.411% -96.371%] (p = 0.00 < 0.05)
                        thrpt:  [+2655.4% +2686.4% +2719.3%]
                        Performance has improved.
```

### Gain:

20,000 insertions in BTree/FST:
- Before: 1.16 s 
- After:  78 ms

20,000 insertions in BTree/Trie:
- Before: 2.16 s 
- After:  78 ms

Following this outcome, `index_bench` is now set to 100_000 insertions.

## Is this related to any issues?

No

## Have you read the [Contributing Guidelines](https://github.com/surrealdb/surrealdb/blob/main/CONTRIBUTING.md)?

- [x] I have read the [Contributing Guidelines](https://github.com/surrealdb/surrealdb/blob/main/CONTRIBUTING.md)